pyproject.toml
[project]
name = "diffusionsr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.1.0",
    "pillow>=12.0.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
